Update and rename hearders/to_string.h to headers/to_string.h

+++ headers/to_string.h
@@ -1,6 +1,6 @@
 #ifndef __TO_STRING__
 #define __TO_STRING__
 
-string to_string(int number);
+string to_string(double number);
 
 #endif
